@@ -1,6 +1,6 @@
 {
   "name": "vuex",
-  "version": "2.4.0",
+  "version": "3.0.0",
   "description": "state management for Vue.js",
   "main": "dist/vuex.common.js",
   "module": "dist/vuex.esm.js",
@@ -25,7 +25,7 @@
     "test:ssr": "rollup -c build/rollup.dev.config.js && VUE_ENV=server jasmine JASMINE_CONFIG_PATH=test/unit/jasmine.json",
     "test:types": "tsc -p types/test",
     "release": "bash build/release.sh",
-    "docs": "cd docs && gitbook serve",
+    "docs": "cd docs && gitbook install && gitbook serve",
     "docs:deploy": "cd docs && ./deploy.sh"
   },
   "repository": {
@@ -40,34 +40,38 @@
   "homepage": "https://github.com/vuejs/vuex#readme",
   "devDependencies": {
     "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.10",
+    "babel-loader": "^7.1.2",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-polyfill": "^6.22.0",
     "babel-preset-env": "^1.5.1",
-    "chromedriver": "^2.27.2",
+    "chromedriver": "^2.32.3",
     "cross-spawn": "^5.0.1",
-    "css-loader": "^0.26.1",
-    "eslint": "^3.15.0",
+    "css-loader": "^0.28.7",
+    "eslint": "^3.19.0",
     "eslint-plugin-vue-libs": "^1.2.0",
     "express": "^4.14.1",
-    "jasmine": "2.5.3",
-    "jasmine-core": "2.5.2",
+    "gitbook-plugin-edit-link": "^2.0.2",
+    "gitbook-plugin-github": "^3.0.0",
+    "gitbook-plugin-prism": "^2.3.0",
+    "gitbook-plugin-theme-vuejs": "^1.1.0",
+    "jasmine": "2.8.0",
+    "jasmine-core": "2.8.0",
     "nightwatch": "^0.9.12",
     "nightwatch-helpers": "^1.2.0",
     "phantomjs-prebuilt": "^2.1.14",
-    "rollup": "^0.41.4",
-    "rollup-plugin-buble": "^0.15.0",
-    "rollup-plugin-replace": "^1.1.1",
-    "rollup-watch": "^3.2.2",
+    "rollup": "^0.50.0",
+    "rollup-plugin-buble": "^0.16.0",
+    "rollup-plugin-replace": "^2.0.0",
+    "rollup-watch": "^4.3.1",
     "selenium-server": "^2.53.1",
     "todomvc-app-css": "^2.1.0",
-    "typescript": "^2.1.5",
-    "uglify-js": "^2.7.5",
-    "vue": "^2.4.3",
-    "vue-loader": "^12.0.0",
-    "vue-template-compiler": "^2.4.3",
-    "webpack": "^2.6.1",
+    "typescript": "^2.5.3",
+    "uglify-js": "^3.1.2",
+    "vue": "github:vuejs/vue#dev",
+    "vue-loader": "^13.0.5",
+    "vue-template-compiler": "^2.4.4",
+    "webpack": "^3.6.0",
     "webpack-dev-middleware": "^1.10.0",
-    "webpack-hot-middleware": "^2.16.1"
+    "webpack-hot-middleware": "^2.19.1"
   }
 }
